do not forcibly add cast when inserting table
